+++ assets/js/login.js
@@ -1,0 +1,3 @@
+global.$ = global.jQuery = require('jquery');
+
+require('semantic-ui-css');