+++ webpack.config.js
@@ -13,10 +13,17 @@ Encore
     .addEntry('js/app', [
         './assets/js/app.js'
     ])
+    .addEntry('js/login', [
+        './assets/js/login.js'
+    ])
     .addStyleEntry('css/app', [
         './assets/css/app.css',
         './vendor/friendsofsymfony/comment-bundle/Resources/public/css/comments.css'
     ])
+    .addStyleEntry('css/login', [
+        './assets/css/login.css',
+    ])
+
 ;
 
 module.exports = Encore.getWebpackConfig();